third_party/WebKit/LayoutTests/dom/text/text-should-not-have-attributes-expected.txt

This tests that Text nodes do no have attributes nor hasAttributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Text.prototype.hasOwnProperty("hasAttributes") is false
PASS "hasAttributes" in text is false
PASS "attributes" in text is false
PASS successfullyParsed is true

TEST COMPLETE
